moved configuration so seperate file in server dir
serverconf.sh has to be present in server dir.
server.sh can be moved to a folder in $PATH to
manage multiple servers without needing to manage multiple instances of
the script.

A seperate tmux socket is created for every sever instance to avoid
errors with multiple running servers

added option to choose backup method in serverconf.sh